Palindrome phrase length check
This simple program check the lenght of a phrase and if that phrase is palindrome sets a bit in the stack to 1, differently from the other program this is made with the official motorola syntax that i havent used in precedent program because I dont like it and I found it very displeasing to  use in any case this is how it should be written if we had an official motorola processor. this program still isn't optimized but is a rather first approach made when i was tired might get new implementation if needed.